feat: 딜러 도메인 생성
