+++ src/test/java/blackjack/domain/DealerTest.java
@@ -83,35 +83,3 @@ class DealerTest {
         assertThat(dealer.getScore()).isEqualTo(15);
     }
 }
-
-class Dealer {
-
-    private static final int MAXIMUM_CARD_COUNT = 2;
-    private static final int MAXIMUM_SCORE = 16;
-
-    private final Cards cards;
-
-    public Dealer(final Cards cards) {
-        this.cards = cards;
-    }
-
-    public void hit(final Card card) {
-        cards.addCard(card);
-    }
-
-    public boolean isHittable() {
-        return isCardShortage() && isScoreLow();
-    }
-
-    private boolean isCardShortage() {
-        return cards.count() <= MAXIMUM_CARD_COUNT;
-    }
-
-    private boolean isScoreLow() {
-        return cards.calculateTotalScore() <= MAXIMUM_SCORE;
-    }
-
-    public int getScore() {
-        return cards.calculateTotalScore();
-    }
-}
